Move the ToolLibraryEditor code to its own file
